refactor(#8/post): 컨트롤러 패키지 이름을 presentation으로 변경
